pyproject.toml
[project]
name = "vizora"
version = "0.1.0"
description = ""
authors = [
    {name = "Brandon-Nicholson",email = "brandon.s.nicholson@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "openai (>=2.14.0,<3.0.0)",
    "numpy (>=2.4.0,<3.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "rapidfuzz (>=3.14.3,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "typer (>=0.21.1,<0.22.0)",
    "rich (>=14.2.0,<15.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "python-multipart (>=0.0.12,<1.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "supabase (>=2.0.0,<3.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "reportlab (>=4.0.0,<5.0.0)",
    "pillow (>=10.0.0,<11.0.0)",
    "stripe (>=7.0.0,<8.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "apscheduler (>=3.10.0,<4.0.0)",
    "resend (>=2.0.0,<3.0.0)",
    "google-api-python-client (>=2.100.0,<3.0.0)",
    "google-auth-oauthlib (>=1.0.0,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
